+++ baybe/kernels/composite.py
@@ -4,7 +4,8 @@ import gc
 from functools import reduce
 from operator import add, mul
 
-from attrs import define, field
+import numpy as np
+from attrs import Attribute, cmp_using, define, field
 from attrs.converters import optional as optional_c
 from attrs.validators import deep_iterable, gt, instance_of, min_len
 from attrs.validators import optional as optional_v
@@ -81,6 +82,46 @@ class ProductKernel(CompositeKernel):
     @override
     def to_gpytorch(self, *args, **kwargs):
         return reduce(mul, (k.to_gpytorch(*args, **kwargs) for k in self.base_kernels))
+
+
+@define(frozen=True)
+class ProjectionKernel(CompositeKernel):
+    """A projection kernel for dimensionality reduction."""
+
+    base_kernel: Kernel = field(validator=instance_of(Kernel))
+    """The kernel applied to the projected inputs."""
+
+    projection_matrix: np.ndarray = field(
+        converter=np.asarray, eq=cmp_using(eq=np.array_equal)
+    )
+    """The projection matrix."""
+
+    learn_projection: bool = field(
+        default=True, validator=instance_of(bool), kw_only=True
+    )
+    """Boolean specifying if the projection matrix should be learned.
+
+    If ``True``, the provided projection matrix is used as initial value."""
+
+    @projection_matrix.validator
+    def _validate_projection_matrix(self, _: Attribute, value: np.ndarray):
+        if value.ndim != 2:
+            raise ValueError(
+                f"The projection matrix must be 2-dimensional, "
+                f"but has shape {value.shape}."
+            )
+
+    @override
+    def to_gpytorch(self, **kwargs):
+        from baybe.kernels._gpytorch import GPyTorchProjectionKernel
+
+        n_projections = self.projection_matrix.shape[-1]
+        gpytorch_kernel = self.base_kernel.to_gpytorch(ard_num_dims=n_projections)
+        return GPyTorchProjectionKernel(
+            gpytorch_kernel,
+            projection_matrix=self.projection_matrix,
+            learn_projection=self.learn_projection,
+        )
 
 
 # Collect leftover original slotted classes processed by `attrs.define`